Solve ProductDetails circular dependencies
